+++ src/main/java/org/opentripplanner/routing/algorithm/transferoptimization/configure/TransferOptimizationServiceConfigurator.java
@@ -128,11 +128,12 @@ public class TransferOptimizationServiceConfigurator<T extends RaptorTripSchedul
   private DefaultCostCalculator<T> createCostCalculator() {
     McCostParams p = raptorRequest.multiCriteriaCostFactors();
     return new DefaultCostCalculator<>(
-        transitDataProvider.stopBoarAlightCost(),
         p.boardCost(),
         p.transferCost(),
         p.walkReluctanceFactor(),
-        p.waitReluctanceFactor()
+        p.waitReluctanceFactor(),
+        transitDataProvider.stopBoarAlightCost(),
+        p.transitReluctanceFactors()
     );
   }
 }